Separate out card and container, and further Notification styling
