+++ client/src/App.js
@@ -1,4 +1,4 @@
-import React, { useState, useEffect } from 'react'
+import React, { useState, useEffect, useContext } from 'react'
 
 import {
   BrowserRouter as Router,
@@ -6,14 +6,18 @@ import {
   Route,
   Redirect
 } from 'react-router-dom'
+import { css } from '@emotion/react'
 
 import Home from './components/Home'
 import Loading from './components/Loading'
 import LoginForm from './components/LoginForm'
 import SignupForm from './components/SignupForm'
-import Notification from './components/Notification'
+import Notification, { ERROR } from './components/Notification'
 import Layout from './components/layout'
-import Container from './components/Container'
+import Container from './components/layout/Container'
+import Card from './components/layout/Card'
+import { setNotification } from './reducer'
+import { store } from './store'
 
 const App = () => {
   const [token, setToken] = useState(null)
@@ -39,27 +43,35 @@ const App = () => {
     }
   }
 
+  const { dispatch } = useContext(store)
+
   return (
     <Router>
       <Layout>
-        <Notification />
         <Container>
-          <Switch>
-            <Route path="/login">
-              <LoginForm updateToken={updateToken} />
-            </Route>
+          <Notification
+            css={css`
+              margin-bottom: 0.75em;
+            `}
+          />
+          <Card>
+            <Switch>
+              <Route path="/login">
+                <LoginForm updateToken={updateToken} />
+              </Route>
 
-            <Route path="/signup">
-              <SignupForm></SignupForm>
-            </Route>
+              <Route path="/signup">
+                <SignupForm></SignupForm>
+              </Route>
 
-            <Route path="/">
-              <Loading loading={loading}>
-                {token ? <Home /> : <Redirect to="/login" />}
-              </Loading>
-            </Route>
-          </Switch>
-          {token && <button onClick={() => logout()}>Logout</button>}
+              <Route path="/">
+                <Loading loading={loading}>
+                  {token ? <Home /> : <Redirect to="/login" />}
+                </Loading>
+              </Route>
+            </Switch>
+            {token && <button onClick={() => logout()}>Logout</button>}
+          </Card>
         </Container>
       </Layout>
     </Router>
