--- client/src/components/Container.js
@@ -1,40 +1,0 @@
-import React, { useContext } from 'react'
-import PropTypes from 'prop-types'
-import { css } from '@emotion/react'
-
-import { store } from '../store'
-
-const Container = ({ children }) => {
-  const { state } = useContext(store)
-
-  const { theme } = state
-
-  return (
-    <div
-      css={css`
-        height: 100vh;
-        display: flex;
-        align-items: center;
-      `}
-    >
-      <div
-        css={css`
-          width: 100%;
-          margin: 0 auto;
-          max-width: 400px;
-          background: ${theme.colors.background3};
-          border-radius: 0.5em;
-          padding: 1.5em;
-        `}
-      >
-        {children}
-      </div>
-    </div>
-  )
-}
-
-Container.propTypes = {
-  children: PropTypes.any
-}
-
-export default Container
